+++ libraries/WaveMix/WaveMix.cpp
@@ -1,8 +1,8 @@
 //
 //    FILE: WaveMix.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.0
-// PURPOSE: WaveMix library for Arduino
+// VERSION: 0.1.1
+// PURPOSE: Arduino library to mix two signals (A and B) with an adaptive weight.
 //     URL: https://github.com/RobTillaart/WaveMix
 //
 
@@ -12,36 +12,34 @@
 
 WaveMix::WaveMix()
 {
-  _w1 = 0.5;
-  _w2 = 0.5;
+  _weight[0] = 0.5;
+  _weight[1] = 0.5;
 }
 
 
-void WaveMix::setWeight(float w1, float w2)
+void WaveMix::setWeight(float weight1, float weight2)
 {
-  float t = 1.0 / (w1 + w2);
-  _w1 = w1 * t;
-  _w2 = w2 * t;
+  float factor = 1.0 / (weight1 + weight2);
+  _weight[0] = weight1 * factor;
+  _weight[1] = weight2 * factor;
 }
 
 
-void WaveMix::setPercentage(float p)
+void WaveMix::setPercentage(float percentage)
 {
-  if (p > 100)    p = 100;
-  else if (p < 0) p = 0;
-  _w1 = p * 0.01;
-  _w2 = 1.0 - _w1;
+  _weight[0] = percentage * 0.01;
+  _weight[1] = 1.0 - _weight[0];
 }
 
 
 float WaveMix::mix(float s1, float s2)
 {
-  // are these optimizations efficient?
-  if (_w1 == 0) return s2;
-  if (_w2 == 0) return s1;
-  return _w1 * s1 + _w2 * s2;
-  
+  if (_gain == 0) return 0;
+  if (_weight[0] == 0) return s2 * _gain;
+  if (_weight[1] == 0) return s1 * _gain;
+  return ((_weight[0] * s1) + (_weight[1] * s2)) * _gain;
 }
 
 
 // -- END OF FILE --
+