+++ libraries/WaveMix/WaveMix.h
@@ -2,15 +2,15 @@
 //
 //    FILE: WaveMix.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.0
-// PURPOSE: an Arduino library to mix two signals (A and B) with an adaptive weight.
-//     URL: https://github.com/RobTillaart/CountDown
+// VERSION: 0.1.1
+// PURPOSE: Arduino library to mix two signals (A and B) with an adaptive weight.
+//     URL: https://github.com/RobTillaart/WaveMix
 //
 
 
 #include "Arduino.h"
 
-#define WAVEMIX_LIB_VERSION               (F("0.1.0"))
+#define WAVEMIX_LIB_VERSION               (F("0.1.1"))
 
 
 class WaveMix
@@ -18,19 +18,23 @@ class WaveMix
 public:
   explicit WaveMix();
 
-  //  w1 >= 0 && w2 >= 0 && (w1+w2) > 0
-  void  setWeight(float w1, float w2);
-  float getW1() { return _w1; };
-  float getW2() { return _w2; };
-  //  0 <= p <= 100
-  void  setPercentage(float p);
+  //  weight1 + weight2 != 0
+  void  setWeight(float weight1, float weight2);
+  float getW1() { return _weight[0]; };
+  float getW2() { return _weight[1]; };
+  //  preferably 0 <= percentage <= 100
+  void  setPercentage(float percentage);
 
+  void  setGain(float gain) { _gain = gain; };
+  float getGain() { return _gain; };
 
-  float mix(float s1, float s2);
+  float mix(float s1, float s2 = 0);
+
 
 private:
-  float _w1 = 0.5;
-  float _w2 = 0.5;
+  float _weight[2] = { 0.5, 0.5 };
+  float _gain      = 1.0;
+
 };