+++ libraries/WaveMix/CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2022-03-26
+
+### Added
+- add **setGain()** and **getGain()**
+- add examples
+
+### Changed
+- updated readme.md
+- removed constraints on weight and percentage
+
+### Fixed
+- minor edits and clean up 
+
+
 ## [0.1.0] - 2022-03-24
 
 ### Added